backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.3.0
pydantic==2.5.2
google-generativeai==0.3.1
python-dotenv==1.0.0
python-multipart==0.0.6
pdfplumber==0.10.3
requests==2.31.0
